consolidate log and trace processor
