+++ src/OpenTelemetry/Logs/LogProcessor.cs
@@ -15,169 +15,17 @@
 // </copyright>
 
 #if NETSTANDARD2_0
-using System;
-using System.Threading;
-using OpenTelemetry.Internal;
+using OpenTelemetry.Common;
 
 namespace OpenTelemetry.Logs
 {
     /// <summary>
     /// Log processor base class.
     /// </summary>
-    public abstract class LogProcessor : IDisposable
+    public abstract class LogProcessor : BaseProcessor<LogRecord>
     {
-        private int shutdownCount;
-
-        /// <summary>
-        /// Called synchronously when a <see cref="LogRecord"/> is recorded.
-        /// </summary>
-        /// <param name="record">
-        /// The log record.
-        /// </param>
-        /// <remarks>
-        /// This function is called synchronously on the thread which generated
-        /// the log record. This function should be thread-safe, and should not
-        /// block indefinitely or throw exceptions.
-        /// </remarks>
-        public virtual void OnLog(in LogRecord record)
-        {
-        }
-
-        /// <summary>
-        /// Flushes the <see cref="LogProcessor"/>, blocks the current
-        /// thread until flush completed, shutdown signaled or timed out.
-        /// </summary>
-        /// <param name="timeoutMilliseconds">
-        /// The number of milliseconds to wait, or <c>Timeout.Infinite</c> to
-        /// wait indefinitely.
-        /// </param>
-        /// <returns>
-        /// Returns <c>true</c> when flush succeeded; otherwise, <c>false</c>.
-        /// </returns>
-        /// <exception cref="System.ArgumentOutOfRangeException">
-        /// Thrown when the <c>timeoutMilliseconds</c> is smaller than -1.
-        /// </exception>
-        /// <remarks>
-        /// This function guarantees thread-safety.
-        /// </remarks>
-        public bool ForceFlush(int timeoutMilliseconds = Timeout.Infinite)
-        {
-            if (timeoutMilliseconds < 0 && timeoutMilliseconds != Timeout.Infinite)
-            {
-                throw new ArgumentOutOfRangeException(nameof(timeoutMilliseconds));
-            }
-
-            try
-            {
-                return this.OnForceFlush(timeoutMilliseconds);
-            }
-            catch (Exception ex)
-            {
-                // TODO: replace this with a better internal logging
-                OpenTelemetrySdkEventSource.Log.SpanProcessorException(nameof(this.ForceFlush), ex);
-                return false;
-            }
-        }
-
-        /// <summary>
-        /// Attempts to shutdown the <see cref="LogProcessor"/>, blocks
-        /// the current thread until shutdown completed or timed out.
-        /// </summary>
-        /// <param name="timeoutMilliseconds">
-        /// The number of milliseconds to wait, or <c>Timeout.Infinite</c> to
-        /// wait indefinitely.
-        /// </param>
-        /// <returns>
-        /// Returns <c>true</c> when shutdown succeeded; otherwise, <c>false</c>.
-        /// </returns>
-        /// <exception cref="System.ArgumentOutOfRangeException">
-        /// Thrown when the <c>timeoutMilliseconds</c> is smaller than -1.
-        /// </exception>
-        /// <remarks>
-        /// This function guarantees thread-safety. Only the first call will
-        /// win, subsequent calls will be no-op.
-        /// </remarks>
-        public bool Shutdown(int timeoutMilliseconds = Timeout.Infinite)
-        {
-            if (timeoutMilliseconds < 0 && timeoutMilliseconds != Timeout.Infinite)
-            {
-                throw new ArgumentOutOfRangeException(nameof(timeoutMilliseconds));
-            }
-
-            if (Interlocked.Increment(ref this.shutdownCount) > 1)
-            {
-                return false; // shutdown already called
-            }
-
-            try
-            {
-                return this.OnShutdown(timeoutMilliseconds);
-            }
-            catch (Exception ex)
-            {
-                OpenTelemetrySdkEventSource.Log.SpanProcessorException(nameof(this.Shutdown), ex);
-                return false;
-            }
-        }
-
-        /// <inheritdoc/>
-        public void Dispose()
-        {
-            this.Dispose(true);
-            GC.SuppressFinalize(this);
-        }
-
-        /// <summary>
-        /// Called by <c>ForceFlush</c>. This function should block the current
-        /// thread until flush completed, shutdown signaled or timed out.
-        /// </summary>
-        /// <param name="timeoutMilliseconds">
-        /// The number of milliseconds to wait, or <c>Timeout.Infinite</c> to
-        /// wait indefinitely.
-        /// </param>
-        /// <returns>
-        /// Returns <c>true</c> when flush succeeded; otherwise, <c>false</c>.
-        /// </returns>
-        /// <remarks>
-        /// This function is called synchronously on the thread which called
-        /// <c>ForceFlush</c>. This function should be thread-safe, and should
-        /// not throw exceptions.
-        /// </remarks>
-        protected virtual bool OnForceFlush(int timeoutMilliseconds)
-        {
-            return true;
-        }
-
-        /// <summary>
-        /// Called by <c>Shutdown</c>. This function should block the current
-        /// thread until shutdown completed or timed out.
-        /// </summary>
-        /// <param name="timeoutMilliseconds">
-        /// The number of milliseconds to wait, or <c>Timeout.Infinite</c> to
-        /// wait indefinitely.
-        /// </param>
-        /// <returns>
-        /// Returns <c>true</c> when shutdown succeeded; otherwise, <c>false</c>.
-        /// </returns>
-        /// <remarks>
-        /// This function is called synchronously on the thread which made the
-        /// first call to <c>Shutdown</c>. This function should not throw
-        /// exceptions.
-        /// </remarks>
-        protected virtual bool OnShutdown(int timeoutMilliseconds)
-        {
-            return true;
-        }
-
-        /// <summary>
-        /// Releases the unmanaged resources used by this class and optionally
-        /// releases the managed resources.
-        /// </summary>
-        /// <param name="disposing">
-        /// <see langword="true"/> to release both managed and unmanaged resources;
-        /// <see langword="false"/> to release only unmanaged resources.
-        /// </param>
-        protected virtual void Dispose(bool disposing)
+        /// <inheritdoc />
+        public sealed override void OnStart(LogRecord record)
         {
         }
     }
